Extract repeated openAuthModal function
