+++ src/Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarArtists.tsx
@@ -8,9 +8,7 @@ import { createFragmentContainer, graphql } from "react-relay"
 
 import { ArtworkSidebarArtists_artwork } from "__generated__/ArtworkSidebarArtists_artwork.graphql"
 import { FollowArtistButtonFragmentContainer as FollowArtistButton } from "Components/FollowButton/FollowArtistButton"
-
-import { stringify } from "qs"
-import { data as sd } from "sharify"
+import { openAuthModal } from "Utils/openAuthModal"
 
 export interface ArtistsProps {
   artwork: ArtworkSidebarArtists_artwork
@@ -32,41 +30,9 @@ export class ArtworkSidebarArtists extends React.Component<ArtistsProps> {
   }
 
   handleOpenAuth = (mediator, artist) => {
-    if (sd.IS_MOBILE) {
-      this.openMobileAuth(artist)
-    } else if (mediator) {
-      this.openDesktopAuth(mediator, artist)
-    } else {
-      window.location.href = "/login"
-    }
-  }
-
-  openMobileAuth = artist => {
-    const params = stringify({
-      action: "follow",
+    openAuthModal(mediator, {
+      entity: artist,
       contextModule: "Artwork page",
-      intent: "follow artist",
-      kind: "artist",
-      objectId: artist.id,
-      signUpIntent: "follow artist",
-      trigger: "click",
-      entityName: artist.name,
-    })
-    const href = `/sign_up?redirect-to=${window.location}&${params}`
-
-    window.location.href = href
-  }
-
-  openDesktopAuth = (mediator, artist) => {
-    mediator.trigger("open:auth", {
-      mode: "signup",
-      copy: `Sign up to follow ${artist.name}`,
-      signupIntent: "follow artist",
-      afterSignUpAction: {
-        kind: "artist",
-        action: "follow",
-        objectId: artist.id,
-      },
     })
   }
